+++ app/src/main/java/com/company/mobile/android/appname/app/bufferoos/viewmodel/BufferoosViewModel.kt
@@ -2,13 +2,13 @@ package com.company.mobile.android.appname.app.bufferoos.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import com.company.mobile.android.appname.app.bufferoos.master.BufferoosNavigationCommand
 import com.company.mobile.android.appname.app.bufferoos.master.BufferoosNavigationCommand.GoToDetailsView
 import com.company.mobile.android.appname.app.common.model.ResourceState
 import com.company.mobile.android.appname.app.common.model.ResourceState.Error
 import com.company.mobile.android.appname.app.common.model.ResourceState.Loading
 import com.company.mobile.android.appname.app.common.model.ResourceState.Success
+import com.company.mobile.android.appname.app.common.viewmodel.CommonEventsViewModel
 import com.company.mobile.android.appname.app.common.viewmodel.SingleLiveEvent
 import com.company.mobile.android.appname.domain.bufferoo.interactor.GetBufferoos
 import com.company.mobile.android.appname.model.bufferoo.Bufferoo
@@ -16,13 +16,13 @@ import io.reactivex.disposables.Disposable
 
 typealias BufferoosState = ResourceState<List<Bufferoo>>
 
-class BufferoosViewModel(private val getBufferoosUseCase: GetBufferoos) : ViewModel() {
+class BufferoosViewModel(private val getBufferoosUseCase: GetBufferoos) : CommonEventsViewModel() {
 
     private val bufferoosLiveData: MutableLiveData<BufferoosState> = MutableLiveData()
     private val selectedBufferooLiveData: MutableLiveData<Bufferoo> = MutableLiveData()
     private var bufferoos: List<Bufferoo> = emptyList()
     private var disposable: Disposable? = null
-    val bufferoosNavigationEvent = SingleLiveEvent<BufferoosNavigationCommand>()
+    val bufferoosNavigationLiveEvent = SingleLiveEvent<BufferoosNavigationCommand>()
 
     override fun onCleared() {
         disposable?.dispose()
@@ -42,7 +42,7 @@ class BufferoosViewModel(private val getBufferoosUseCase: GetBufferoos) : ViewMo
         // Store selected item for details view
         selectedBufferooLiveData.value = bufferoos[position]
         // Go to details view
-        bufferoosNavigationEvent.value = GoToDetailsView
+        bufferoosNavigationLiveEvent.value = GoToDetailsView
     }
 
     fun fetchBufferoos() {
@@ -50,11 +50,17 @@ class BufferoosViewModel(private val getBufferoosUseCase: GetBufferoos) : ViewMo
         // the view just after setting it.
         bufferoosLiveData.value = Loading()
         disposable = getBufferoosUseCase.execute()
-            .subscribe({
-                bufferoos = it
-                bufferoosLiveData.value = Success(bufferoos)
-            }, {
-                bufferoosLiveData.value = Error(it.message ?: "")
-            })
+            .subscribeWith(
+                object : SingleRemoteInterceptor<List<Bufferoo>>(commonLiveEvent) {
+                    override fun onSuccess(bufferoos: List<Bufferoo>) {
+                        this@BufferoosViewModel.bufferoos = bufferoos
+                        bufferoosLiveData.value = Success(bufferoos)
+                    }
+
+                    override fun onRegularError(e: Throwable) {
+                        bufferoosLiveData.value = Error(e.message ?: "")
+                    }
+                }
+            )
     }
 }