+++ datasources/src/main/java/com/company/mobile/android/appname/datasources/remote/exception/ThrowableWrapperException.kt
@@ -1,0 +1,11 @@
+package com.company.mobile.android.appname.datasources.remote.exception
+
+/**
+ * [Exception] throw by the remote package when a there is a network connection returns a [Throwable].
+ */
+class ThrowableWrapperException : Exception {
+
+    constructor() : super()
+
+    constructor(cause: Throwable) : super(cause)
+}