+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [X.X.X] - YYYY-MM-DD
 ### Added
+- Common events view model was implemented. The first common event is the unauthorized error of remote requests.
 - Form validators with styles were added.
 - Sign in activity was implemented.
 - A basic example with no navigation was included, too.